ci: Also check types with Pyrefly
