+++ pyproject.toml
@@ -96,6 +96,7 @@ samples = [
 typing = [
   "backports-strenum",
   "mypy>=1.9",
+  "pyrefly>=0.60",
   "pyright>=1.1.408",
   "sphinx",
   "ty>=0.0.1a15",
@@ -272,6 +273,13 @@ warn_unused_ignores = true
 overrides = [ { ignore_missing_imports = true, module = [
   "nox.*",
 ] } ]
+
+[tool.pyrefly]
+search-path = [ "." ]
+[[tool.pyrefly.sub-config]]
+matches = "tests/**"
+[tool.pyrefly.sub-config.errors]
+no-access = false
 
 [tool.ty]
 [tool.ty.rules]
